Extract out types into packet.go
