+++ x/ibcratelimit/keeper/rate_limit.go
@@ -14,42 +14,6 @@ import (
 	"github.com/provenance-io/provenance/x/ibcratelimit/types"
 )
 
-var (
-	msgSend = "send_packet"
-	msgRecv = "recv_packet"
-)
-
-type UndoSendMsg struct {
-	UndoSend UndoPacketMsg `json:"undo_send"`
-}
-
-type UndoPacketMsg struct {
-	Packet UnwrappedPacket `json:"packet"`
-}
-
-type SendPacketMsg struct {
-	SendPacket PacketMsg `json:"send_packet"`
-}
-
-type RecvPacketMsg struct {
-	RecvPacket PacketMsg `json:"recv_packet"`
-}
-
-type PacketMsg struct {
-	Packet UnwrappedPacket `json:"packet"`
-}
-
-type UnwrappedPacket struct {
-	Sequence           uint64                                `json:"sequence"`
-	SourcePort         string                                `json:"source_port"`
-	SourceChannel      string                                `json:"source_channel"`
-	DestinationPort    string                                `json:"destination_port"`
-	DestinationChannel string                                `json:"destination_channel"`
-	Data               transfertypes.FungibleTokenPacketData `json:"data"`
-	TimeoutHeight      clienttypes.Height                    `json:"timeout_height"`
-	TimeoutTimestamp   uint64                                `json:"timeout_timestamp,omitempty"`
-}
-
 func (k Keeper) CheckAndUpdateRateLimits(ctx sdk.Context, msgType string, packet exported.PacketI) error {
 	contract := k.GetContractAddress(ctx)
 
@@ -86,7 +50,7 @@ func (k Keeper) UndoSendRateLimit(ctx sdk.Context, contract string, packet expor
 		return err
 	}
 
-	msg := UndoSendMsg{UndoSend: UndoPacketMsg{Packet: unwrapped}}
+	msg := types.UndoSendMsg{UndoSend: types.UndoPacketMsg{Packet: unwrapped}}
 	asJSON, err := json.Marshal(msg)
 	if err != nil {
 		return err
@@ -100,17 +64,17 @@ func (k Keeper) UndoSendRateLimit(ctx sdk.Context, contract string, packet expor
 	return nil
 }
 
-func (k Keeper) unwrapPacket(packet exported.PacketI) (UnwrappedPacket, error) {
+func (k Keeper) unwrapPacket(packet exported.PacketI) (types.UnwrappedPacket, error) {
 	var packetData transfertypes.FungibleTokenPacketData
 	err := json.Unmarshal(packet.GetData(), &packetData)
 	if err != nil {
-		return UnwrappedPacket{}, err
+		return types.UnwrappedPacket{}, err
 	}
 	height, ok := packet.GetTimeoutHeight().(clienttypes.Height)
 	if !ok {
-		return UnwrappedPacket{}, types.ErrBadMessage
+		return types.UnwrappedPacket{}, types.ErrBadMessage
 	}
-	return UnwrappedPacket{
+	return types.UnwrappedPacket{
 		Sequence:           packet.GetSequence(),
 		SourcePort:         packet.GetSourcePort(),
 		SourceChannel:      packet.GetSourceChannel(),
@@ -130,13 +94,13 @@ func (k Keeper) BuildWasmExecMsg(msgType string, packet exported.PacketI) ([]byt
 
 	var asJSON []byte
 	switch {
-	case msgType == msgSend:
-		msg := SendPacketMsg{SendPacket: PacketMsg{
+	case msgType == types.MsgSendPacket:
+		msg := types.SendPacketMsg{SendPacket: types.PacketMsg{
 			Packet: unwrapped,
 		}}
 		asJSON, err = json.Marshal(msg)
-	case msgType == msgRecv:
-		msg := RecvPacketMsg{RecvPacket: PacketMsg{
+	case msgType == types.MsgRecvPacket:
+		msg := types.RecvPacketMsg{RecvPacket: types.PacketMsg{
 			Packet: unwrapped,
 		}}
 		asJSON, err = json.Marshal(msg)
@@ -156,11 +120,12 @@ func (k Keeper) RevertSentPacket(
 	ctx sdk.Context,
 	packet exported.PacketI,
 ) error {
-	contract := k.GetContractAddress(ctx)
-	if contract == "" {
+	if !k.ContractConfigured(ctx) {
 		// The contract has not been configured. Continue as usual
 		return nil
 	}
+
+	contract := k.GetContractAddress(ctx)
 
 	return k.UndoSendRateLimit(
 		ctx,
