+++ x/ibcratelimit/types/packet.go
@@ -1,0 +1,42 @@
+package types
+
+import (
+	transfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
+	clienttypes "github.com/cosmos/ibc-go/v6/modules/core/02-client/types"
+)
+
+const (
+	MsgSendPacket = "send_packet"
+	MsgRecvPacket = "recv_packet"
+)
+
+type UndoSendMsg struct {
+	UndoSend UndoPacketMsg `json:"undo_send"`
+}
+
+type UndoPacketMsg struct {
+	Packet UnwrappedPacket `json:"packet"`
+}
+
+type SendPacketMsg struct {
+	SendPacket PacketMsg `json:"send_packet"`
+}
+
+type RecvPacketMsg struct {
+	RecvPacket PacketMsg `json:"recv_packet"`
+}
+
+type PacketMsg struct {
+	Packet UnwrappedPacket `json:"packet"`
+}
+
+type UnwrappedPacket struct {
+	Sequence           uint64                                `json:"sequence"`
+	SourcePort         string                                `json:"source_port"`
+	SourceChannel      string                                `json:"source_channel"`
+	DestinationPort    string                                `json:"destination_port"`
+	DestinationChannel string                                `json:"destination_channel"`
+	Data               transfertypes.FungibleTokenPacketData `json:"data"`
+	TimeoutHeight      clienttypes.Height                    `json:"timeout_height"`
+	TimeoutTimestamp   uint64                                `json:"timeout_timestamp,omitempty"`
+}
